pyproject.toml
[tool.poetry]
name = "blackjack-backend"
version = "0.1.0"
description = ""
authors = ["mackmc22 <mackmc22@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
django-rest-framework = "^0.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
